implemented repository

+++ app/src/main/java/com/gruppa/books/data/db/entities/BookOrderCountEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 
 @Entity(tableName = "order_books_count", primaryKeys = ["orderId", "bookId"])
 data class BookOrderCountEntity(
-    val orderId: Int,
+    val orderId: Long,
     val bookId: Long,
     val count: Int,
 )
